fix: move location of PendingTxs files
