+++ src/components/Dashboard/PendingTxs/PendingTxsList/index.tsx
@@ -9,7 +9,7 @@ import { sm } from 'src/theme/variables'
 import { currentChainId } from 'src/logic/config/store/selectors'
 import { generateSafeRoute, SAFE_ROUTES } from 'src/routes/routes'
 import { getChainById } from 'src/config'
-import PendingTxListItem from 'src/components/Dashboard/PendingTxListItem'
+import PendingTxListItem from 'src/components/Dashboard/PendingTxs/PendingTxListItem'
 import { currentSafe } from 'src/logic/safe/store/selectors'
 import { MAX_TXS_DISPLAY } from 'src/routes/Home'
 import { useQueueTransactions } from 'src/routes/safe/components/Transactions/TxList/hooks/useQueueTransactions'
